Orbital laser to old MEDFORD implementation to clean repo
